+++ test/all.test.js
@@ -632,6 +632,37 @@ describe(Support.getTestDialectTeaser('Tests'), function () {
 						expect(folder.children[0].children[0].children[1].children).not.to.exist;
 					});
 				});
+
+				it('works with `raw` option', function() {
+					return this.folder.findAll({order: [['name']], hierarchy: true, raw: true}).bind(this)
+					.then(function(folders) {
+						expect(folders.length).to.equal(1);
+						var folder = folders[0];
+
+						expect(folder.name).to.equal('a');
+						expect(folder.children).to.exist;
+						expect(folder.children.length).to.equal(2);
+						expect(folder.children[0].name).to.equal('ab');
+						expect(folder.children[1].name).to.equal('ac');
+
+						expect(folder.children[0].children).to.exist;
+						expect(folder.children[0].children.length).to.equal(2);
+						expect(folder.children[0].children[0].name).to.equal('abd');
+						expect(folder.children[0].children[1].name).to.equal('abe');
+
+						expect(folder.children[1].children).not.to.exist;
+
+						expect(folder.children[0].children[0].children).to.exist;
+						expect(folder.children[0].children[0].children.length).to.equal(2);
+						expect(folder.children[0].children[0].children[0].name).to.equal('abdf');
+						expect(folder.children[0].children[0].children[1].name).to.equal('abdg');
+
+						expect(folder.children[0].children[1].children).not.to.exist;
+
+						expect(folder.children[0].children[0].children[0].children).not.to.exist;
+						expect(folder.children[0].children[0].children[1].children).not.to.exist;
+					});
+				});
 			});
 
 			describe('works with scoped models', function () {